fix: discount 메소드 반환타입 Long으로 변경

+++ src/main/java/christmas/domain/event/discountEvent/D_DayEvent.java
@@ -1,6 +1,6 @@
 package christmas.domain.event.discountEvent;
 
-public enum D_DayEvent implements Event<Integer, Integer, Integer>{
+public enum D_DayEvent implements Event<Integer, Long, Integer>{
 
     DISCOUNT_DAY(EventConfig.START_DATE, EventConfig.END_DATE);
 
@@ -18,7 +18,7 @@ public enum D_DayEvent implements Event<Integer, Integer, Integer>{
     }
 
     @Override
-    public Integer discount(Integer date) {
+    public Long discount(Integer date) {
         return EventConfig.INITAL_DISCOUNT_AMOUNT + EventConfig.DAILY_DISCOUNT_INCREASE * date;
     }
 }
